Normaliza modelos del dominio para alinearse con esquema MySQL
- ConsumoLote: cambia List<Producto> a Producto único (coincide con FK id_producto)
- Producto: precio → precioUnidad, stock(int) → stockActual(double)
- Proveedor: celular → telefono (coincide con columna BD)
- Cantidad en DetalleCompra/Pedido/Venta, ConsumoLote y MovimientoAlmacen: int → double (DECIMAL BD)
- DetallePedido: subTotal → subtotal (consistencia con otros detalles)
- MovimientoAlmacen: java.util.Date → java.sql.Date, descripcion → contexto
- Pedido: java.util.Date → java.sql.Date, campos ahora private
- Agrega constructores vacíos donde faltaban

+++ src/main/java/Dominio/Modelo/DetalleCompra.java
@@ -4,13 +4,13 @@ public class DetalleCompra {
     private int idDetalle;
     private Compra compra;
     private Producto producto;
-    private int cantidad;
+    private double cantidad;
     private double subtotal;
 
     public DetalleCompra() {
     }
 
-    public DetalleCompra(int idDetalle, Compra compra, Producto producto, int cantidad, double subtotal) {
+    public DetalleCompra(int idDetalle, Compra compra, Producto producto, double cantidad, double subtotal) {
         this.idDetalle = idDetalle;
         this.compra = compra;
         this.producto = producto;
@@ -42,11 +42,11 @@ public class DetalleCompra {
         this.producto = producto;
     }
 
-    public int getCantidad() {
+    public double getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
